+++ crates/vfs/src/lib.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 //! VFS assistance for moss including optimised tree + blit helpers
-pub use self::tree::Tree;
+pub use self::tree::{BlitFile, Tree, builder::TreeBuilder};
 
 pub mod cache;
 pub mod path;
